added new data type, new controls for user first signin
