+++ prisma/schema.prisma
@@ -65,6 +65,15 @@ enum Verification {
   REJECTED
 }
 
+enum Occupation {
+  Azubi
+  Bachelor
+  Master
+  Doctor
+  Professor
+  Researcher
+}
+
 model User {
   id            String       @id @default(cuid()) @map("_id") // @db.ObjectId
   name          String?
@@ -79,7 +88,7 @@ model User {
   verification  Verification @default(UNVERIFIED)
   updated       Boolean      @default(false)
   birthDate     DateTime?
-  occupation    String?
+  occupation    Occupation?
   location      String?
   affiliation   String[]     @default([]) // change to normal string instead of list
   createdAt     DateTime     @default(now())
