+++ src/server/trpc/router/userRouter.ts
@@ -5,7 +5,7 @@ import { router } from "../trpc";
 import { getSession } from "./user/getSession";
 import { getUser } from "./user/getUser";
 import { getUsers } from "./user/getUsers";
-
+import { updateUserUpdateProfile } from "./user/updateUserUpdateProfile";
 export const userRouter = router({
   getSession,
   getUser,
@@ -13,4 +13,5 @@ export const userRouter = router({
   getUsers,
   getUserById,
   updateUserById,
+  updateUserUpdateProfile,
 });
